Updates install instructions and code cleanup

+++ inst/include/samplers.hpp
@@ -12,7 +12,7 @@ public:
   Sampler(/* args */);
   ~Sampler();
 
-  static void metropolis_hastings(unsigned *a_t, Eigen::VectorXd *w_t, int N, unsigned t, int B = 10);
+  static void metropolis_hastings(unsigned *a_t, Eigen::VectorXd *w_t, size_t N, int t, size_t B = 10);
 
   // Add other possible samplers here:
   // void gibbs();
